cart validation done

+++ src/components/CartPage/CartProduct.jsx
@@ -9,7 +9,20 @@ const CartProduct = ({ product }) => {
   const dispatch = useDispatch();
 
   const handleDelete = () => {
-    dispatch(deleteProductFromCartThunk(product.id))
+    Swal.fire({
+      title: `Are you sure on deleting ${product.product.title} ?`,
+      text: "You won't be able to revert this!",
+      icon: "warning",
+      showCancelButton: true,
+      confirmButtonColor: "#3085d6",
+      cancelButtonColor: "#d33",
+      confirmButtonText: "Yes, delete it!"
+    }).then((result) => {
+      if (result.isConfirmed) {
+        dispatch(deleteProductFromCartThunk(product.id))
+      }
+    });
+    
   }
 
   return (
